Fix binary jail build, more mount work
The first problem with the binary jail build is that the binaries were
put in the wrong directory, and because of that were not tar'd.
The second problem is several directories weren't created (src, tmp,
obj), and the third problem is that no src files exist in the binary
releases.  I don't know if the third problem is actually a problem yet
because the pkgsrc builds shouldn't need the source at all.

For the mounts, changes were made to recognize and handle HAMMER FS
differently.  Trying to mount NFS on a hammer partition will screw up
the machine, you must use NULLFS instead.