Fix DragonFly Jail environment variables
The DragonFly release ISOs have no source code, so first we need to get
param.h from /usr/include rather than the src directory.  Secondly,
DragonFly has a different format for BRANCH and REVISION, so the awk
code was adjusted.  For the ISOs, the update.sh will have to be updated
to create a faked "newvers.sh" file in the otherwise empty src folder.