Create new tc command: getDependencyCascades
While I can't shake the feeling that I'm missing something, I just have not
been able to identify where run/lib dependencies packages of dependencies
would get pulled in on FreeBSD.  It's certainly not happening on pkgsrc.

In an attempt to fix this, I've created a new tc command,
getDependencyCascades, which will return all run/lib dependencies of
a port's dependencies.

The output of this command will be sent to portbuild so it can pull in the
needed packages.