Cascading mystery solved
It turns out that my intuition was right.  The problem the whole time was
that Buildlink3 needs to be considered a depends for both build and run,
and I was only considering it as a build dependency.

By adding a single line to the makefile fixed the issue.

The "getDependencyCascades" is not used internally, but it's still a nice
addition.  Well, I think anyway...