Fixed first-time FTP fetch and leftover file problem.
Previously the very first fetch of a distfile would result in an
ftp fetch failure after a successful fetch (file was going in
/tmp/distfiles and it was directed to look in /distcache as the
distcache option was enabled.)  Also, the distfile that got dumped
in distcache would then show as a leftover file (a file that wasn't
there before the first pristine snapshot).

We don't want this file removed (it's supposed to be cached) so I
just added /distcache to the exclusion list to fix it.