Support multiple jobs
FreeBSD by default uses multiple jobs, and you have to disable it with
MAKE_JOBS_UNSAFE on each port.  Pkgsrc uses the same MAKE_JOBS_UNSAFE but
also requires the number of jobs to be explicitly requested.

We'll use the same jobs = cpus*2 + 1 formula used on make world.
We'll also use the same global env var to disable it.