Implemented per-distfile options capability
This seemed easy at the start, but several hours later and many small
problems is all it took to realize what a big job the options file is.

I suspect I'm not yet done, I need to check if the makemake functionality
brings in files based on the options.  I think probably not.