Duds work, mark as "arch" failure
I got boolean-confused with the duds list, the last commit marked all the
good packages as duds.  Since the only real reason for a dud is a bad
architecture (normally not for *-DragonFly but sometimes not for x86_64)
the reason is always marked as "arch" rather than leave the script to
guess.