Remove LFTP requirement; stick with fetch
Since there's no more directory mirroring required, we don't really need
LFTP anymore.  I thought it was just quiet, but apparently the "eval"
command silenced it.

Never the less, LFTP is a dependency that's not needed as /usr/bin/fetch
works great for our purposes, so let's still with that.

Eval was removed, echoes were added to update.logs so something would be
created, plus it's nice to document the actual source of the iso image.

README update accordingly.