Finally fix tinderd startup problem
I finally figured out that the tinderd really didn't have a path to
bmake (just like the error message said), so by modifying the
subshell environment to add /usr/pkg/bin to the path fixed the issue.

FreeBSD doesn't have this issue because their make is on a system
bin directory.