Git world build: add pkgsrc bootstrap
Building world from source has the additional challenge of also building
the pkgsrc bootstrap tools from source.  These aren't part of the base
distribution, but rather the pkgsrc tree.

The solution is to pull the files in from the host computers pkgsrc tree,
but this assumes that is located at /usr/pkg.  For now it's hardcoded in.
Initial review of the resultant jail tarball shows everything is working
fine.