Add skeleton for tinderbox-sunos
It's important not to have separate patch.maps for each version of
tinderbox, so I created a framework for tinderbox-sunos (in practical
terms it is tinderbox-openindiana) and set up the patch generator
how it should be.

This is all the work that I'm doing on tinderbox-sunos in the near term.