Move nullfs to be the default mount
NFS continues to be problematic, and FreeeBSD t/b users say that nullfs is
more suited for the default, so we'll make that happen here.  I had a
problem getting NFS to work on a i386 Dfly (didn't like the -alldirs on a
non-FS mount).

This change greatly simplifies the mount logic, but I lose the nice
hammer detection scheme that I derived a few days ago.

Instructions adjusted accordingly.
If basic regression testing goes well on banshee, this will probably be the
first version of tinderbox-dragonfly that I commit.