Merge BOOTSTRAP_DEPENDS with FETCH_DEPENDS
I don't really want to ask Tinderbox project to change their database
schemas to support BOOTSTRAP_DEPENDS, so I'm changing it to be considered
as a FETCH DEPENDS, and modifying the buildscript accordingly.  This
will reduce the patch count by two and only leave one SQL patch (for
the conftest log marking).  It will also reduce differences between
FreeBSD and this version of Tinderbox.