Move leftover files after package deinstall as info only
It is possible for a database file like mimeinfo.cache to be created
automatically during the installation of a package, but it'll be cleaned
out when the support packaged are deinstalled.  Don't mark the port as
"Leftovers" if the leftover files are gone when all the packages are
removed, but leave the message for informational purposes.