Subtle pkgsrc-specific bug finally found
Somehow the php-* packages spontaneously change their name from php5-*
to something like php53-*.  How this was happening is not understood
because the name came from the same source independently.

The solution is just pass along the pkgname from the portbuild script to
the buildscript script, and now it's working.

This was the cause for the infinite loop, but I'm leaving the safeguard
in place because it doesn't hurt and who knows?