Modify MakeCache module to handle -OR- Perl dependencies
Pkgsrc handles dependencies a little differently than FreeBSD in that
they are conditional dependencies, depending on what's already installed.
In many cases, the dependency does NOT have to be installed if Perl > X
version is already on the system, presumably due to modules becoming
part of Perl.

Therefore, if for these types of dependencies (enclosed in curly brackets)
start with "{perl>=?5" then lang/perl5 will be used, otherwise the
indicated dependency will be used.  This fixes the neverending loop
issue seen (rarely) on Perl packages.