Improve bulk hook: address wiping out queue
The bulk abruptly stopped this morning and the queue was empty.  It
turns out that a directly apparently ends in a trailing space (how is
that possible?) and thus the sed pattern didn't match the second
slash (which was never foreseen as possible).  Adding a "g" modifier
and also checking that the value of SPORTDIR is not empty should
repeat issues in the future.