Temporary fix of pkgsrc bootstrap stage
The new mk/compiler/gcc.mk file now sucks in lang/gcc47/version.mk
so lang/gcc47 has to be added to the list.  The problem is that
this will change over time (adding or replacing it with gcc48) so
the real solution may be to copy over a stored version of gcc.mk
that doesn't include external references to ../../lang.